apps/processor/requirements.txt
httpx == 1.0.0b0
pedalboard==0.5.10
python_dotenv == 0.20.0
rich == 12.5.1
sentry-sdk==1.9.7
git+https://github.com/jps-aldridge/py-simple-audio@master#egg=simpleaudio
SoundFile==0.10.3.post1
wheel==0.37.1